requirements.txt
certifi==2020.12.5
chardet==3.0.4
Django==2.2.13
djangorestframework==3.9.2
h5py==3.2.1
idna==2.10
joblib==1.0.1
jsonfield==3.1.0
Keras==2.4.3
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.24.0
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.25.11
